requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
openpyxl>=3.0.0
